pyproject.toml
[tool.poetry]
name = "stripe-test"
version = "0.1.0"
description = ""
authors = ["Lana <lan2828@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
stripe = "^8.4.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
